requirements.txt
asgiref==3.10.0
Django==5.2.8
pillow==12.0.0
sqlparse==0.5.3
tzdata==2025.2
django-allauth==65.3.0
PyJWT==2.10.1
cryptography==44.0.0
